Split utils file in multiple specific files
