+++ typescript/common-resolvers/src/image/get-image-file-key.ts
@@ -1,0 +1,8 @@
+import mime from "mime-types";
+
+export const getImageFileKey = (
+  imageId: string,
+  workspaceId: string,
+  datasetId: string,
+  mimetype: string
+) => `${workspaceId}/${datasetId}/${imageId}.${mime.extension(mimetype)}`;
