+++ typescript/common-resolvers/src/image/get-image-name.ts
@@ -1,0 +1,18 @@
+export const getImageName = ({
+  externalUrl,
+  finalUrl,
+  name,
+}: {
+  externalUrl?: string | null;
+  finalUrl?: string | null;
+  name?: string | null;
+}): string => {
+  const nameBase =
+    name ??
+    externalUrl?.substring(
+      externalUrl?.lastIndexOf("/") + 1,
+      externalUrl?.indexOf("?")
+    ) ??
+    finalUrl!.substring(finalUrl!.lastIndexOf("/") + 1, finalUrl!.indexOf("?"));
+  return nameBase.replace(/\.[^/.]+$/, "");
+};
