+++ typescript/common-resolvers/src/image/import-and-process-image.ts
@@ -2,11 +2,9 @@ import { v4 as uuidv4 } from "uuid";
 import { ImageCreateInput } from "@labelflow/graphql-types";
 import { Repository, DbImageCreateInput } from "../types";
 import { getOrigin } from "../utils/get-origin";
-import {
-  getImageFileKey,
-  getImageName,
-  throwIfInvalidImageInputs,
-} from "./utils";
+import { validImageInput } from "./validate-image-input";
+import { getImageName } from "./get-image-name";
+import { getImageFileKey } from "./get-image-file-key";
 
 /**
  * Downloads the image from the external url and upload it
@@ -109,7 +107,7 @@ const importImageIfNeeded = async (
   },
   { req, repository }: { req: Request; repository: Repository }
 ): Promise<string> => {
-  throwIfInvalidImageInputs({ file, externalUrl, url });
+  validImageInput({ file, externalUrl, url });
 
   const getImageFileKeyFromMimeType = (mimeType: string) =>
     getImageFileKey(imageId, workspaceId, datasetId, mimeType);
